+++ tests/integration/mod.rs
@@ -1,4 +1,5 @@
 mod creds;
+mod message_retry;
 mod pub_properties;
 mod pub_sub;
 mod session;